+++ analyzers/dataframe/ReconstructedParticle2Track.h
@@ -17,7 +17,7 @@ namespace ReconstructedParticle2Track{
 					   ROOT::VecOps::RVec<edm4hep::TrackState> tracks);
 
   /// Return the charge of a track to a reconstructed particle
-  ROOT::VecOps::RVec<int> getRP2TRK_charge(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in,  
+  ROOT::VecOps::RVec<float> getRP2TRK_charge(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in,  
 					   ROOT::VecOps::RVec<edm4hep::TrackState> tracks);
 
   /// Return the D0 of a track to a reconstructed particle
